requirements.txt
Flask==3.0.3
flask-cors
gunicorn==23.0.0
Werkzeug==3.0.3
celery
requests
numpy
pyyaml
chess
backoff
--extra-index-url https://download.pytorch.org/whl/cpu
torch
